18-APP_FULL/requirements.txt
fastapi[all]==0.73.0
uvicorn
gunicorn
pydantic
asyncpg
sqlalchemy
redis==4.3.3
fastapi-cache2[redis]==0.1.8
prometheus-fastapi-instrumentator==5.9.1
python-logging-loki==0.3.1
opentelemetry-instrumentation-fastapi==0.34b0
opentelemetry-instrumentation-logging==0.34b0
opentelemetry-instrumentation-sqlalchemy==0.34b0
opentelemetry-instrumentation-redis==0.34b0
opentelemetry.instrumentation.aio_pika==0.34b0
opentelemetry.instrumentation.aiohttp_client==0.34b0
opentelemetry-exporter-otlp==1.13.0
opentelemetry-api==1.13.0
opentelemetry-sdk==1.13.0
opentelemetry-exporter-jaeger-thrift==1.13.0
aiohttp